+++ lib/SILOptimizer/Utils/Differentiation/CMakeLists.txt
@@ -3,5 +3,6 @@ silopt_register_sources(
   Common.cpp
   DerivativeLookup.cpp
   DifferentiationInvoker.cpp
+  JVPEmitter.cpp
   LinearMapInfo.cpp
 )
